feat: implementation pkg packages (strings random & time)
